+++ lib/gotrue.dart
@@ -2,9 +2,13 @@ library gotrue;
 
 export 'src/auth_options.dart';
 export 'src/constants.dart' show AuthChangeEvent;
+export 'src/cookie_options.dart';
+export 'src/gotrue_api.dart';
 export 'src/gotrue_client.dart';
+export 'src/gotrue_error.dart';
 export 'src/gotrue_response.dart';
 export 'src/provider.dart';
 export 'src/session.dart';
+export 'src/subscription.dart' show Subscription;
 export 'src/user.dart';
 export 'src/user_attributes.dart';