+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## [next]
+
+- fix: fetch the user, if missing, on `/verify` ([#29](https://github.com/supabase-community/gotrue-dart/issues/29))
+
+## [0.1.5]
+
+- feat: add `toString` method to `GotrueError`class
+
+## [0.1.4]
+
+- fix: trigger signedIn event on recoverSession
+
+## [0.1.3]
+
+- feat: add `tokenRefreshed` auth event
+- feat: add slack, spotify and twitch Auth providers
+- fix: update currentSession.user when GoTrueClient.update is called
+- chore: export missing types
+
 ## [0.1.2]
 
 - feat: `setAuth()` method for setting the session with a provided jwt